+++ ECS/Src/Entities/entity_factory.cpp
@@ -39,7 +39,7 @@ Entity EntityFactory::createPlayer(
     PositionComponent startPosition(0, 0);
     VelocityComponent velocity{100.0f};
     SpriteDataComponent spriteData{SpritePath::Ship1, {16, 40}, {96, 48}, {1.0f, 1.0f}};
-    HitboxComponent hitbox{(float)spriteData.dimension.x, (float)spriteData.dimension.y};
+    HitboxComponent hitbox{spriteData.dimension.x, spriteData.dimension.y};
     HealthComponent health{100, 100};
     InputComponent input{InputType::NONE};
